Issue-2121 move filter method to FilenameUtils
