+++ src/main/java/run/halo/app/service/impl/BackupServiceImpl.java
@@ -106,6 +106,7 @@ import run.halo.app.service.UserService;
 import run.halo.app.utils.DateTimeUtils;
 import run.halo.app.utils.DateUtils;
 import run.halo.app.utils.FileUtils;
+import run.halo.app.utils.FilenameUtils;
 import run.halo.app.utils.HaloUtils;
 import run.halo.app.utils.JsonUtils;
 import run.halo.app.utils.VersionUtil;
@@ -631,7 +632,7 @@ public class BackupServiceImpl implements BackupService {
             content.append(postMarkdownVo.getOriginalContent());
             try {
                 String fileName = postMarkdownVo.getTitle() + "-" + postMarkdownVo.getSlug();
-                String markdownFileName = FileUtils.filterReservedCharsInFileName(fileName)+ ".md";
+                String markdownFileName = FilenameUtils.filterReservedCharsInFileName(fileName)+ ".md";
                 Path markdownFilePath = Paths.get(markdownFileTempPathName, markdownFileName);
                 if (!Files.exists(markdownFilePath.getParent())) {
                     Files.createDirectories(markdownFilePath.getParent());
