+++ src/main/java/run/halo/app/utils/FileUtils.java
@@ -587,9 +587,4 @@ public class FileUtils {
     public static void writeStringToFile(File file, String content) throws IOException {
         org.apache.commons.io.FileUtils.writeStringToFile(file, content, StandardCharsets.UTF_8);
     }
-
-    public static String filterReservedCharsInFileName(String fileName){
-        Pattern filePattern = Pattern.compile("[\\\\/:*?\"<>|]");
-        return filePattern.matcher(fileName).replaceAll("");
-    }
 }
