+++ src/test/java/run/halo/app/utils/FilenameUtilsTest.java
@@ -45,4 +45,15 @@ class FilenameUtilsTest {
         assertEquals("tar.gz", FilenameUtils.getExtension("he/ll/o.tar.gz"));
         assertEquals("tar.bz2", FilenameUtils.getExtension("he/ll/o.tar.bz2"));
     }
+
+    @Test
+    public void fileNameWithReservedCharsWillBeReplaced(){
+        String fileName1 = "abcde";
+        String filteredFileName1 = FilenameUtils.filterReservedCharsInFileName(fileName1) + ".md";
+        assertEquals("abcde.md", filteredFileName1);
+
+        String fileName2 = "abcde|字符替换\\星号*大于>小于<slash/中文字符、";
+        String filteredFileName2 = FilenameUtils.filterReservedCharsInFileName(fileName2) + ".md";
+        assertEquals("abcde字符替换星号大于小于slash中文字符、.md", filteredFileName2);
+    }
 }
